pyproject.toml
[tool.poetry]
name = "language-overdrive"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "language_overdrive"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.2.0"
google-cloud-translate = "^3.9.0"
openai = "^0.25.0"
python-dotenv = "^0.21.0"
fastapi = {extras = ["all"], version = "^0.88.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
